requirements.txt
moviepy>=2.0
Pillow>=10.0
numpy>=1.24
elevenlabs>=1.0
pytest>=7.0
python-dotenv>=1.0
python-telegram-bot>=20.0
